# apps/backend/auth-service/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AuthResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"attributes": {
								"items": {},
								"type": "array"
							},
							"IdToken": {
								"type": "string"
							},
							"RefreshToken": {
								"type": "string"
							},
							"AccessToken": {
								"type": "string"
							},
							"email": {
								"type": "string"
							}
						},
						"type": "object",
						"nullable": true
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"createAuthRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"verifyAuthRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"verificationCode": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"verificationCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"signInAuthRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserAttributesResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"attributes": {
								"items": {},
								"type": "array"
							}
						},
						"required": [
							"attributes"
						],
						"type": "object"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"email": {
						"type": "string"
					},
					"cognitoId": {
						"type": "string"
					},
					"isEmailVerified": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"email",
					"cognitoId",
					"isEmailVerified",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IUser"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UsersResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"currentPage": {
								"type": "number",
								"format": "double"
							},
							"totalUsers": {
								"type": "number",
								"format": "double"
							},
							"users": {
								"items": {
									"$ref": "#/components/schemas/IUser"
								},
								"type": "array"
							}
						},
						"required": [
							"totalPages",
							"currentPage",
							"totalUsers",
							"users"
						],
						"type": "object"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "auth-service",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/createAuthRequest"
							}
						}
					}
				}
			}
		},
		"/auth/verify": {
			"post": {
				"operationId": "Verify",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/verifyAuthRequest"
							}
						}
					}
				}
			}
		},
		"/auth/signin": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/signInAuthRequest"
							}
						}
					}
				}
			}
		},
		"/auth/google/login": {
			"get": {
				"operationId": "LoginWithGoogle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"data",
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/google/callback": {
			"get": {
				"operationId": "CognitoCallback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										},
										"status": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"data",
										"message",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/user/{email}/attributes": {
			"get": {
				"operationId": "GetUserAttributes",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserAttributesResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/user/{email}": {
			"get": {
				"operationId": "GetUserByEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsersResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}